+++ packages/@biomejs/backend-jsonrpc/src/workspace.ts
@@ -913,7 +913,7 @@ export interface Nursery {
 	 */
 	noDuplicateTestHooks?: RuleConfiguration_for_Null;
 	/**
-	 * Disallow variables from evolving into any type through reassignments. In TypeScript, variables without explicit type annotations can evolve their types based on subsequent assignments. This behavior can inadvertently lead to variables with an any type, weakening type safety. Just like the any type, evolved any types disable many type checking rules and should be avoided to maintain strong type safety. This rule prevents such cases by ensuring variables do not evolve into any type, encouraging explicit type annotations and controlled type evolutions.
+	 * Disallow variables from evolving into any type through reassignments.
 	 */
 	noEvolvingAny?: RuleConfiguration_for_Null;
 	/**